Refactor navigation and auth components structure
- Move Header and SideMenu to navigation components
- Update import paths in protected layout
- Remove deprecated AuthLayout
- Add optional onSignout callback to SignoutButton
- Update Login screen import for AuthLayout
