+++ feature/auth/components/index.tsx
@@ -1,12 +1,9 @@
-// Form Inputs
+// Actions
+export { SignoutButton } from "./actions/SignoutButton";
+
+// Form
 export { EmailInput } from "./form/EmailInput";
 export { PasswordInput } from "./form/PasswordInput";
 
-// Layout
-export { AuthLayout } from "./layout/AuthLayout";
-
-// Guards
+// Guard
 export { AuthRedirectGuard } from "./guard/AuthRedirectGuard";
-
-// Actions
-export { SignoutButton } from "./actions/SignoutButton";
